node to page: Handle aliases, 404, robots.txt, sitemap
Updates #2297
